refactor: improve error handling, testing, and production readiness
- CI/CD: Update GitHub Actions workflows for code coverage and security
  audits, switch to XPlat coverage format and lcov.info for clients
- Client: Add comprehensive test coverage with vitest coverage-v8, improve
  API error handling with fallback to statusText, disable auto-open in
  Docker, add meta tags to HTML
- Server: Implement standardized ErrorResponse type, add CORS policy with
  configurable origins, improve DatabaseApiClient with robust error
  handling and URL escaping, add validation attributes and production config
- Tests: Fix floating point precision in metrics tests
- Docs: Update paths (hpc/ → hpc_runner/), correct terminology (controller
  → .NET API server), clarify API proxy configuration, update CI references
  to GitHub Actions
- Tooling: Update justfile and package.json to reflect client rename and
  remove obsolete Python controller references

This refactoring enhances production readiness with better error handling,
CORS configuration, and comprehensive test coverage while improving
documentation accuracy.

+++ server.core/Domain/ProcessingMetric.cs
@@ -19,8 +19,11 @@ public class ProcessingMetric
     [JsonPropertyName("pdfPages")]
     public int? PdfPages { get; set; }
 
+    /// <summary>
+    /// PDF file size in bytes. Uses long to support files >2GB.
+    /// </summary>
     [JsonPropertyName("pdfSizeBytes")]
-    public int? PdfSizeBytes { get; set; }
+    public long? PdfSizeBytes { get; set; }
 
     [JsonPropertyName("success")]
     public bool Success { get; set; }
@@ -28,6 +31,9 @@ public class ProcessingMetric
     [JsonPropertyName("errorMessage")]
     public string? ErrorMessage { get; set; }
 
+    /// <summary>
+    /// Timestamp of metric creation. Should be stored and handled as UTC.
+    /// </summary>
     [JsonPropertyName("createdAt")]
     public DateTime CreatedAt { get; set; }
 }
